feat: Integrate React Query for global data fetching and refactor contributor data management using a new hook and provider.
