+++ package.json
@@ -38,8 +38,10 @@
 		"@radix-ui/react-tabs": "^1.1.13",
 		"@radix-ui/react-toggle": "^1.1.10",
 		"@radix-ui/react-tooltip": "^1.2.8",
+		"@tanstack/react-query": "^5.90.12",
 		"@types/d3": "^7.4.3",
 		"@types/mdx": "^2.0.13",
+		"axios": "^1.13.2",
 		"class-variance-authority": "^0.7.1",
 		"clsx": "^2.1.1",
 		"cmdk": "^1.1.1",
